requirements.txt
PyYAML
influxdb
